Create Disambiguated_master_company_list.py
The overall problem is around standardizing the data set- using a bunch of techniques around disambiguation of base entities. Note the use of data-structures to reduce the "working/search/contextual space", and tricks around flattening a list. I use concepts from Divide & conquer & Greedy Approach problem solving in this context, along with tricks on handling duplicates (for example -treating lists as sets- which do not have duplicates by definition, and thus can be used to narrow down the work-space further). 

Keep It Simple, Stupid !

----SPECIFICS ---
The prelim step involves looking up the data to do frequency plots on raw data - this gives an idea of the transformations to do on it, including filtering for stop words. This is important since we do Frequent item set mining - and don't want to give more weightage to stop words like the in "The Bank of America" , The Bank of New York Mellon" - This, of course is only possible since we had a good look at the frequecy maps of the raw data-set 
ALGORITHM
1. Transforming the data, Handling any known abbrevations(example -[24]7 inc is the same as 247)converting to lower case, removing the stop words - This will create a list of similar companies 
[Key is to think about reducing the working space, so that it is "scan'nable by human eyes" for discrepancies]
2. Do Frequency item set mining on the transformed data-set
3. Plot the Pretty table on the "standardized data-set"